fix: Now retrieve auto generated id of workspace upon creation to return to front end for subsequent api calls.

+++ packages/api/src/main/java/com/nsa/bt/scrumble/models/Workspace.java
@@ -3,13 +3,13 @@ package com.nsa.bt.scrumble.models;
 import java.util.ArrayList;
 
 public class Workspace {
-    private int id;
+    private long id;
     private User createdBy;
     private String name;
     private String description;
     private ArrayList<Integer> projectIds;
 
-    public Workspace(int id, User createdBy, String name, String description, ArrayList<Integer> projectIds) {
+    public Workspace(long id, User createdBy, String name, String description, ArrayList<Integer> projectIds) {
         this.id = id;
         this.createdBy = createdBy;
         this.name = name;
@@ -28,11 +28,11 @@ public class Workspace {
         this.projectIds = projectIds;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
